refactor: extracted code into smaller components
